change component name showcase => showlist + creation of a component with array from API
